+++ Engine/ac/drawingsurface.cpp
@@ -299,7 +299,7 @@ void DrawingSurface_DrawStringWrapped(ScriptDrawingSurface *sds, int xx, int yy,
     for (size_t i = 0; i < Lines.Count(); i++)
     {
         GUI::DrawTextAlignedHor(ds, Lines[i], font, text_color, sds->GetBlendMode(),
-            xx, xx + wid - 1, yy + linespacing*i, (FrameAlignment)alignment);
+            xx, xx + wid - 1, yy + linespacing * static_cast<int>(i), (FrameAlignment)alignment);
     }
 
     sds->FinishedDrawing();
